Moved sample code to separate module.
